+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
 
+### Added
+- feat: ignore files from linting using a glob in the config
+
 ## 0.1.1
 - fix: explitily check that file to check is not a directory so directories with file extensions in their name do not crash the linter